Deleted UDPHandler, was unneeded
